requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
faiss-cpu==1.7.4
pypdf==3.17.1
httpx==0.25.2
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
numpy==1.24.3